requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
peewee>=3.17,<4.0
simpleeval>=0.9.13,<1.0
dash>=3.0,<4.0
requests>=2.32,<3.0
gmqtt>=0.7,<1.0
pymodbus>=3.7,<4.0
pyserial>=3.5,<4.0
python-dotenv>=1.0,<2.0

httpx>=0.28,<1.0
eval-type-backport>=0.2,<1.0
